requirements.txt
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.2.0
Requests==2.31.0
streamlit==1.30.0
tensorflow==2.13.0
tensorflow_hub==0.15.0
textblob==0.17.1
